refactor(app): extract NddScoreGeneTable columns, formatters, filter UI (#397)
Split NddScoreGeneTable.vue (1156 -> 898 lines) by extracting the column/
select-option/help catalog (nddScoreGeneTableColumns.ts), the pure cell
formatters and badge-variant helpers (nddScoreGeneTableFormatters.ts), and the
filter-control presentation helpers (nddScoreGeneTableFilterUi.ts). Behavior,
props, network calls, and all NDDScore 'machine-learning / not curated' copy
are preserved verbatim. Adds unit tests for the new pure seams.

Advances #397. Part of #346.
